chore: league create 컴포넌트화
